requirements.txt
aiohttp>=3.7.4.post0
alembic==1.13.0
aniso8601==9.0.1
asgiref==3.3.4
async-timeout==3.0.1
attrs>=22.2.0,<24.0.0

bcrypt==4.1.2
blinker==1.7.0
certifi==2021.5.30
click==8.1.7
cloudinary==1.36.0
distlib==0.3.7
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.1.0
email-validator==1.1.3
exceptiongroup==1.1.3
Faker==22.1.0
filelock==3.13.1
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==1.2.0
Flask-Migrate==4.0.5
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==20.1.0

idna==2.10
importlib_resources==6.3.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema>=4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
multidict==5.1.0
mysql-connector-python==8.0.25
mysqlclient==2.0.3
packaging==24.1
pika==1.3.2
Pillow==8.3.1
pipenv==2023.11.15
platformdirs==3.11.0
pluggy==1.3.0
protobuf==3.17.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycryptodome==3.20.0
PyJWT==2.1.0
PyMySQL==1.0.2
pytest==7.4.3
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==1.0.1
pytz==2021.1
referencing==0.34.0
repoze.lru==0.7
Routes==2.5.1
rpds-py==0.18.0
SQLAlchemy==2.0.23
SQLAlchemy-serializer==1.4.1
sqlparse==0.4.1
tomli==2.0.1
typing-extensions>=4.0.0
virtualenv==20.24.6
waitress==2.0.0
Werkzeug==3.0.1
whitenoise==5.2.0
WTForms==3.1.2
yarl==1.6.3
